+++ releaseNotes.md
@@ -1,3 +1,13 @@
+_March 18th, 2019_  **v1.0.114**  
+- EventStore service(s)
+- Added ReadEventStreamVersionAsync, removed includeEvents from ReadEventStreamAsync.
+
+Thanks,  
+_Slacquer_  
+
+<br>
+<br>
+
 _March 18th, 2019_  **v1.0.113**  
 - AssemblyResolver, found a bug where it would return null rather than a cached assembly.
 - EventStore exceptions, renamed.